requirements.txt
langchain>=0.1.0
langgraph>=0.0.10
chromadb>=0.4.22
python-dotenv>=1.0.0
langchain_community>=0.3.19
sentence-transformers>=2.2.2
transformers>=4.37.2
torch>=2.2.0
langchain-core>=0.1.0
pypdf>=3.15.1
requests>=2.31.0